+++ src/documents/item.mjs
@@ -137,5 +137,4 @@ export class CastleFalkensteinItem extends Item {
      // Post message to chat
     CastleFalkenstein.createChatMessage(this.actor, flavor, content);
   }
-
 }